BaSyx Schema Conversion & Performance Improvements (#178)
* changed schema

* fixed infrastructure to new schema

* fixed errors

* simplified filter query logic

* new version


* fix

* fix lint

* refactor

* fix

* Intermediate Commit with new Submodel Repository Database Code


* Implements other Endpoints

* Changes

* adapts schema

* Fixes IT

* Finalizes Conversion

* Adapts launch.json back

* Fixes test failures

* Adds & fixes missing Tests

* Fixes Performance issue

* Code Refactor

* Update internal/common/model/model_value_only_converter.go


---------

+++ cmd/digitaltwinregistryservice/main.go
@@ -89,7 +89,8 @@ func runServer(ctx context.Context, configPath string, databaseSchema string) er
 
 	registryDatabase, err := registrydb.NewPostgreSQLAASRegistryDatabase(
 		dsn,
-		cfg.Postgres.MaxOpenConnections,
+		//nolint:gosec // configured value is bounded by deployment configuration
+		int32(cfg.Postgres.MaxOpenConnections),
 		cfg.Postgres.MaxIdleConnections,
 		cfg.Postgres.ConnMaxLifetimeMinutes,
 		cfg.Server.CacheEnabled,
@@ -102,7 +103,8 @@ func runServer(ctx context.Context, configPath string, databaseSchema string) er
 
 	discoveryDatabase, err := discoverydb.NewPostgreSQLDiscoveryBackend(
 		dsn,
-		cfg.Postgres.MaxOpenConnections,
+		//nolint:gosec // configured value is bounded by deployment configuration
+		int32(cfg.Postgres.MaxOpenConnections),
 		cfg.Postgres.MaxIdleConnections,
 		cfg.Postgres.ConnMaxLifetimeMinutes,
 		databaseSchema,
