+++ cmd/discoveryservice/main.go
@@ -86,7 +86,8 @@ func runServer(ctx context.Context, configPath string, databaseSchema string) er
 
 	smDatabase, err := persistencepostgresql.NewPostgreSQLDiscoveryBackend(
 		dsn,
-		cfg.Postgres.MaxOpenConnections,
+		//nolint:gosec // configured value is bounded by deployment configuration
+		int32(cfg.Postgres.MaxOpenConnections),
 		cfg.Postgres.MaxIdleConnections,
 		cfg.Postgres.ConnMaxLifetimeMinutes,
 		databaseSchema,
